# server/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: History Wiki API
  version: 0.1.0
  description: 历史 Wiki 的公开读者 API。空数据库行为是本契约的一等状态。
  license:
    name: Unlicensed
servers:
  - url: /
security: []
paths:
...
  /api/v1/admin/event-imports/preflight:
    post:
      operationId: preflightEventImport
      summary: 只读预检查批量历史事件草稿
      description: 编辑者和管理员均可使用；返回逐条字段及业务错误，不创建幂等记录、审计、历史事件、活动草稿或批次记录。请求最多 10,000 条且 JSON 请求体最多 20 MiB。
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventImportRequest" }
      responses:
        "200":
          description: 预检查结果；存在错误时仍返回 200 且 valid=false
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventImportValidation" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-history_wiki_session
  parameters:
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: 必须与 Secure CSRF Cookie 一致，并与当前服务端 Session 绑定。
      schema: { type: string, minLength: 1 }
...
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    EventImportRecord:
      type: object
      additionalProperties: false
      required: [id, slug]
      properties:
        id:
          type: string
          format: uuid
          pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-7[0-9a-fA-F]{3}-[89aAbB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
          description: 调用方提供的不可变 UUIDv7；导入不会自动替换标识。
        slug: { type: string, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$', maxLength: 120 }
        title: { type: string, maxLength: 200, default: "" }
        summary: { type: string, maxLength: 1000, default: "" }
        narrative: { type: string, maxLength: 100000, default: "" }
        time:
          oneOf:
            - $ref: "#/components/schemas/TimeExpression"
            - type: "null"
        primaryCategory:
          oneOf:
            - $ref: "#/components/schemas/PrimaryCategory"
            - type: "null"
        prominence:
          oneOf:
            - $ref: "#/components/schemas/EventProminence"
            - type: "null"
        displayOrder: { type: integer, minimum: -1000000, maximum: 1000000, default: 1000 }
        regionIds: { type: array, maxItems: 20, uniqueItems: true, items: { type: string, format: uuid }, default: [] }
        placeIds: { type: array, maxItems: 50, uniqueItems: true, items: { type: string, format: uuid }, default: [] }
        periodIds: { type: array, maxItems: 20, uniqueItems: true, items: { type: string, format: uuid }, default: [] }
        figureIds: { type: array, maxItems: 100, uniqueItems: true, items: { type: string, format: uuid }, default: [] }
        topicTagIds: { type: array, maxItems: 50, uniqueItems: true, items: { type: string, format: uuid }, default: [] }
    EventImportRequest:
      type: object
      additionalProperties: false
      required: [events]
      properties:
        events:
          type: array
          minItems: 1
          maxItems: 10000
          items: { $ref: "#/components/schemas/EventImportRecord" }
    EventImportIssue:
      type: object
      additionalProperties: false
      required: [index, field, code, detail]
      properties:
        index: { type: integer, minimum: 0, description: events 数组中的零基下标 }
        field: { type: string }
        code: { type: string }
        detail: { type: string }
    EventImportValidation:
      type: object
      additionalProperties: false
      required: [valid, total, errors]
      properties:
        valid: { type: boolean }
        total: { type: integer, minimum: 0, maximum: 10000 }
        errors:
          type: array
          items: { $ref: "#/components/schemas/EventImportIssue" }
...
    Problem:
      type: object
      required: [type, title, status, detail, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        requestId: { type: string }
...
    TimeExpression:
      oneOf:
        - $ref: "#/components/schemas/ExactDateExpression"
        - $ref: "#/components/schemas/YearExpression"
        - $ref: "#/components/schemas/CircaExpression"
        - $ref: "#/components/schemas/IntervalExpression"
      discriminator:
        propertyName: kind
    HistoricalYear:
      type: object
      required: [era, year]
      properties:
        era: { type: string, enum: [BCE, CE] }
        year: { type: integer, minimum: 1, maximum: 999999 }
    ExactDateExpression:
      type: object
      required: [kind, date]
      properties:
        kind: { type: string, const: exact-date }
        date:
          allOf:
            - $ref: "#/components/schemas/HistoricalYear"
            - type: object
              required: [month, day]
              properties:
                month: { type: integer, minimum: 1, maximum: 12 }
                day: { type: integer, minimum: 1, maximum: 31 }
    YearExpression:
      type: object
      required: [kind, year]
      properties:
        kind: { type: string, const: year }
        year: { $ref: "#/components/schemas/HistoricalYear" }
    CircaExpression:
      type: object
      required: [kind, year]
      properties:
        kind: { type: string, const: circa }
        year: { $ref: "#/components/schemas/HistoricalYear" }
    IntervalExpression:
      type: object
      required: [kind, start, end]
      properties:
        kind: { type: string, const: interval }
        start:
          allOf:
            - $ref: "#/components/schemas/HistoricalYear"
            - type: object
              properties:
                circa: { type: boolean }
        end:
          allOf:
            - $ref: "#/components/schemas/HistoricalYear"
            - type: object
              properties:
                circa: { type: boolean }
    PrimaryCategory:
      type: string
      enum: [政治, 军事, 文化, 科技, 社会, 交流]
    EventProminence:
      type: integer
      enum: [1, 2, 3]